refactor: Deduplicate isRecord
