+++ packages/@n8n/agents/src/skills/validator.ts
@@ -1,3 +1,4 @@
+import { isRecord } from '@n8n/utils';
 import { parse as parseYaml } from 'yaml';
 
 import {
@@ -484,8 +485,4 @@ function hasContractFields(value: object): boolean {
 
 function optionalString(value: unknown): string | undefined {
 	return typeof value === 'string' && value.trim().length > 0 ? value.trim() : undefined;
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-	return typeof value === 'object' && value !== null && !Array.isArray(value);
 }
