+++ packages/@n8n/instance-ai/evaluations/utils/safe-extract.ts
@@ -1,9 +1,9 @@
 // Type guards for pulling fields off `unknown` records — used wherever we
 // inspect event payloads, run inputs/outputs, or other loosely-typed JSON.
 
-export function isRecord(value: unknown): value is Record<string, unknown> {
-	return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
+import { isRecord } from '@n8n/utils';
+
+export { isRecord };
 
 export function getNestedRecord(
 	obj: Record<string, unknown>,
