+++ packages/@n8n/instance-ai/src/tools/workflows.tool.ts
@@ -4,6 +4,7 @@
  * restore-version, update-version.
  */
 import { Tool } from '@n8n/agents';
+import { isRecord } from '@n8n/utils';
 import type { WorkflowJSON } from '@n8n/workflow-sdk';
 import { nanoid } from 'nanoid';
 import { z } from 'zod';
@@ -695,10 +696,6 @@ async function handleValidate(
 			error: error instanceof Error ? error.message : 'Failed to validate workflow',
 		};
 	}
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-	return typeof value === 'object' && value !== null && !Array.isArray(value);
 }
 
 function isWorkflowJson(value: unknown): value is WorkflowJSON {
