+++ packages/@n8n/mcp-apps/src/server/register-mcp-app-tool.ts
@@ -3,6 +3,7 @@ import type {
 	RegisteredTool,
 	ToolCallback,
 } from '@modelcontextprotocol/sdk/server/mcp.js';
+import { isRecord } from '@n8n/utils';
 import type z from 'zod';
 
 import { RESOURCE_URI_META_KEY } from './constants';
@@ -50,8 +51,4 @@ function normalizeMcpAppToolMeta(meta: Record<string, unknown>): Record<string, 
 	}
 
 	return meta;
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-	return typeof value === 'object' && value !== null && !Array.isArray(value);
 }
