+++ packages/@n8n/mcp-apps/src/utils/guards.ts
@@ -1,3 +1,3 @@
-export function isRecord(value: unknown): value is Record<string, unknown> {
-	return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
+import { isRecord } from '@n8n/utils';
+
+export { isRecord };
