+++ packages/@n8n/utils/src/index.ts
@@ -14,3 +14,4 @@ export * from './placeholder';
 export * from './jwt';
 export * from './scrub-secrets';
 export type * from './types';
+export * from './validation';
