+++ packages/@n8n/workflow-sdk/src/validation/index.ts
@@ -1,3 +1,4 @@
+import { isRecord } from '@n8n/utils';
 import get from 'lodash/get';
 import type { INodeTypes, IConnections as N8nIConnections, IDisplayOptions } from 'n8n-workflow';
 import { mapConnectionsByDestination } from 'n8n-workflow';
@@ -1030,10 +1031,6 @@ function validateOutputUsage(
 			);
 		}
 	}
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-	return typeof value === 'object' && value !== null && !Array.isArray(value);
 }
 
 function getSwitchRulesCount(parameters: Record<string, unknown> | undefined): number {
