+++ packages/cli/src/modules/agents/integrations/integration-message-context.service.ts
@@ -1,4 +1,5 @@
 import { Service } from '@n8n/di';
+import { isRecord } from '@n8n/utils';
 import { jsonParse } from 'n8n-workflow';
 
 import { AgentResourceRepository } from '../repositories/agent-resource.repository';
@@ -72,10 +73,6 @@ export class IntegrationMessageContextService implements IntegrationMessageConte
 			return {};
 		}
 	}
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-	return value !== null && typeof value === 'object' && !Array.isArray(value);
 }
 
 export function isIntegrationMessageContext(value: unknown): value is IntegrationMessageContext {
