+++ packages/cli/src/modules/agents/integrations/slack-app-setup.service.ts
@@ -9,6 +9,7 @@ import type {
 import type { User } from '@n8n/db';
 import { UserRepository } from '@n8n/db';
 import { Service } from '@n8n/di';
+import { isRecord } from '@n8n/utils';
 import { Cipher } from 'n8n-core';
 import { jsonParse } from 'n8n-workflow';
 
@@ -90,10 +91,6 @@ interface SlackAppSetupSession {
 	clientSecret: string;
 	signingSecret: string;
 	redirectUrl: string;
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-	return value !== null && typeof value === 'object' && !Array.isArray(value);
 }
 
 function childRecord(
